moved utils to lib dir
